Update README, clean repo
