+++ ChangeLog.md
@@ -2,6 +2,16 @@
 
 This project's release branch is `master`. This log is written from the perspective of the release branch: when changes hit `master`, they are considered released.
 
+### Version v1.0.0
+
+* *Breaking Change* Reduce dependencies
+* *Breaking Change* Unify features of core libraries and application
+* Introduce instances
+* Make proxies optional
+* Payment channel first API
+* Integrate external commit
+* *Breaking Change* Bump to hydra 0.11.0
+
 ### Version v0.3.0
 
 * Ensure Hydra Node interaction threads currently executing blocking calls when asked to close; close immediately when execution is yielded back to them
