--- common/src/Hydra/NodeTypes.hs
@@ -1,1 +1,0 @@
-
