--- common/src/Hydra/ServerOutput.hs
@@ -1,63 +1,0 @@
--- | 
-
-module Hydra.ServerOutput where
-
-import GHC.Generics
-import Data.Aeson
-
-import Data.Set (Set)
-import Data.Time (UTCTime)
-import Data.Text (Text)
-import Hydra.Types
-import Hydra.ClientInput
-import Hydra.Snapshot (Snapshot)
-
--- Copied and adapted from hydra-poc:hydra-node/src/hydra/ServerOutput.hs
--- Anything that took effort was replaced by Value
-
-data ServerOutput tx
-  = PeerConnected {peer :: NodeId}
-  | PeerDisconnected {peer :: NodeId}
-  | HeadIsInitializing {headId :: HeadId, parties :: Set Party}
-  | Committed {headId :: HeadId, party :: Party, utxo :: UTxOType tx}
-  | HeadIsOpen {headId :: HeadId, utxo :: UTxOType tx}
-  | HeadIsClosed
-      { headId :: HeadId
-      , snapshotNumber :: SnapshotNumber
-      , -- | Nominal deadline until which contest can be submitted and after
-        -- which fanout is possible. NOTE: Use this only for informational
-        -- purpose and wait for 'ReadyToFanout' instead before sending 'Fanout'
-        -- as the ledger of our cardano-node might not have progressed
-        -- sufficiently in time yet and we do not re-submit transactions (yet).
-        contestationDeadline :: UTCTime
-      }
-  | HeadIsContested {headId :: HeadId, snapshotNumber :: SnapshotNumber}
-  | ReadyToFanout {headId :: HeadId}
-  | HeadIsAborted {headId :: HeadId, utxo :: UTxOType tx}
-  | HeadIsFinalized {headId :: HeadId, utxo :: UTxOType tx}
-  | CommandFailed {clientInput :: ClientInput}
-  | -- | Given transaction has been seen as valid in the Head. It is expected to
-    -- eventually be part of a 'SnapshotConfirmed'.
-    TxValid {headId :: HeadId, transaction :: tx}
-  | -- | Given transaction was not not applicable to the given UTxO in time and
-    -- has been dropped.
-    TxInvalid {headId :: HeadId, utxo :: UTxOType tx, transaction :: tx, validationError :: ValidationError}
-  | -- | Given snapshot was confirmed and included transactions can be
-    -- considered final.
-    SnapshotConfirmed
-      { headId :: HeadId
-      , snapshot :: Snapshot tx
-      , signatures :: MultiSignature (Snapshot tx)
-      }
-  | GetUTxOResponse {headId :: HeadId, utxo :: UTxOType tx}
-  | InvalidInput {reason :: String, input :: Text}
-  | -- | A friendly welcome message which tells a client something about the
-    -- node. Currently used for knowing what signing key the server uses (it
-    -- only knows one).
-    Greetings {me :: Party}
-  | PostTxOnChainFailed {postChainTx :: PostChainTx tx, postTxError :: PostTxError tx}
-  | RolledBack
-  deriving (Eq, Show, Generic)
-
-instance (ToJSON tx) => ToJSON (ServerOutput tx)
-instance (FromJSON tx) => FromJSON (ServerOutput tx)
