+++ config/preview/config.json
@@ -1,0 +1,6 @@
+<html><head><title>404 Not Found</title></head><body>
+<h1>Not Found</h1>
+The URL you requested was not found.
+<hr>
+Generated by darkhttpd/1.14 on Tue, 01 Aug 2023 05:27:14 GMT
+</body></html>
