+++ default.nix
@@ -57,26 +57,7 @@ let
             hydra.hydra-node.package.components.exes.hydra-tools
           ];
         });
-
-        backend = haskellLib.overrideCabal super.backend (drv: {
-          librarySystemDepends = (drv.librarySystemDepends or []) ++ [
-            cardano-node.cardano-node
-            cardano-node.cardano-cli
-            hydra.hydra-node.package.components.exes.hydra-node
-            hydra.hydra-node.package.components.exes.hydra-tools
-            pkgs.jq
-            pkgs.coreutils
-            livedoc-devnet-script
-          ];
-        });
       };
     });
-
-  hydra-pay-exe = pkgs.runCommandNoCC "hydra-pay" {} ''
-    mkdir -p $out
-    cp -r ${p.exe}/* $out/
-    mv $out/backend $out/hydra-pay
-  '';
-
 in
-p // {  exe = hydra-pay-exe; inherit cardano-node hydra deps;}
+p // { hydra-pay = p.ghc.hydra-pay; inherit cardano-node hydra deps;}
