+++ hydra-pay-core/hydra-pay-core.cabal
@@ -1,5 +1,5 @@
 name:          hydra-pay-core
-version:       0.3.0
+version:       1.0.0
 cabal-version: >=1.10
 build-type:    Simple
